Downloads/pidl/requirements.txt
streamlit==1.31.0
openai==1.54.0
anthropic==0.18.1
python-dotenv==1.0.0
pandas==2.2.0
plotly==5.18.0
radon==6.0.1
pylint==3.0.3
bandit==1.7.6
autopep8==2.0.4
colorama==0.4.6
google-generativeai==0.3.2
sqlalchemy>=2.0.44
scikit-learn==1.4.0
numpy==1.26.3

